AoC22.Tests/Day6/SignalDecoderTests.cs
using AoC22.Solutions.Day6;

namespace AoC22.Tests.Day6;

public class SignalDecoderTests
{
    [Theory]
    [InlineData("mjqjpqmgbljsphdztnvjfqwrcgsmlb", 7)]
    [InlineData("bvwbjplbgvbhsrlpgdmjqwftvncz", 5)]
    [InlineData("nppdvjthqldpwncqszvftbrmjlhg", 6)]
    [InlineData("nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg", 10)]
    [InlineData("zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw", 11)]
    public void FirstPacketCanBeLocated(string input, int expectedFirstPacket)
    {
        var decoder = new SignalDecoder(input);
        var result = decoder.FindFirstPacket();

        result.Should().Be(expectedFirstPacket);
    }
    
    [Theory]
    [InlineData("mjqjpqmgbljsphdztnvjfqwrcgsmlb", 19)]
    [InlineData("bvwbjplbgvbhsrlpgdmjqwftvncz", 23)]
    [InlineData("nppdvjthqldpwncqszvftbrmjlhg", 23)]
    [InlineData("nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg", 29)]
    [InlineData("zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw", 26)]
    public void MessageCanBeLocated(string input, int expectedMessage)
    {
        var decoder = new SignalDecoder(input);
        var result = decoder.FindMessage();

        result.Should().Be(expectedMessage);
    }
}